requirements.txt
# requirements.txt

streamlit
pandas
plotly
Pillow
openpyxl
matplotlib
numpy